requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.6
    # via kombu
apipkg==1.5
    # via execnet
appdirs==1.4.4
    # via
    #   black
    #   fs
asgiref==3.3.4
    # via django
attrs==21.1.0
    # via
    #   interrogate
    #   pytest
billiard==3.6.4.0
    # via celery
black==21.5b0
    # via -r requirements.in
boto3==1.17.68
    # via django-bakery
botocore==1.20.68
    # via
    #   boto3
    #   s3transfer
bump2version==1.0.1
    # via bumpversion
bumpversion==0.6.0
    # via -r requirements.in
celery==5.0.5
    # via -r requirements.in
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   interrogate
    #   pip-tools
colorama==0.4.4
    # via interrogate
coverage==5.5
    # via pytest-cov
dj-database-url==0.5.0
    # via environs
dj-email-url==1.0.2
    # via environs
django-bakery==0.12.7
    # via -r requirements.in
django-cache-url==3.2.3
    # via environs
django-db-geventpool==4.0.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.3.0
    # via revsys-teams
django-health-check==3.16.4
    # via -r requirements.in
django-rest-auth==0.9.5
    # via -r requirements.in
django-test-plus==1.4.0
    # via -r requirements.in
django-tracer==0.9.3
    # via -r requirements.in
django==3.2.2
    # via
    #   -r requirements.in
    #   django-db-geventpool
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-rest-auth
    #   djangorestframework
    #   revsys-teams
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   django-rest-auth
    #   revsys-teams
environs[django]==9.3.2
    # via -r requirements.in
execnet==1.8.0
    # via pytest-xdist
factory-boy==3.2.0
    # via revsys-teams
faker==8.1.2
    # via factory-boy
fs==2.4.13
    # via django-bakery
gevent==21.1.2
    # via -r requirements.in
greenlet==1.1.0
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
interrogate==1.3.2
    # via -r requirements.in
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
kombu==5.0.2
    # via celery
marshmallow==3.11.1
    # via environs
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.18
    # via click-repl
psycogreen==1.0.2
    # via django-db-geventpool
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.10.0
    # via
    #   interrogate
    #   pytest
    #   pytest-forked
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements.in
pytest-django==4.2.0
    # via -r requirements.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-xdist==2.2.1
    # via -r requirements.in
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-xdist
python-dateutil==2.8.1
    # via
    #   botocore
    #   faker
python-dotenv==0.17.1
    # via environs
python-json-logger==2.0.1
    # via -r requirements.in
pytz==2021.1
    # via
    #   celery
    #   django
    #   fs
redis==3.5.3
    # via -r requirements.in
regex==2021.4.4
    # via black
https://afcec3a300c9919dcefedb87c15ecfe09869a9b5@github.com/revsys/revsys-teams/archive/master.zip
    # via -r requirements.in
s3transfer==0.4.2
    # via boto3
shortid==0.1.2
    # via revsys-teams
six==1.16.0
    # via
    #   click-repl
    #   django-bakery
    #   django-rest-auth
    #   fs
    #   python-dateutil
sqlparse==0.4.1
    # via django
structlog==21.1.0
    # via -r requirements.in
tabulate==0.8.9
    # via interrogate
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   interrogate
    #   pep517
    #   pytest
urllib3==1.26.4
    # via botocore
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==5.2.0
    # via -r requirements.in
wrapt==1.12.1
    # via revsys-teams
zope.event==4.5.0
    # via gevent
zope.interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
